send header to order detail component
